Add a safety net for breaking APIs
