+++ .github/workflows/api-compatibility.yml
@@ -1,0 +1,47 @@
+name: API Compatibility
+
+on:
+  pull_request:
+    branches: [ '*' ]
+  push:
+    branches: [ '*' ]
+
+jobs:
+  generate-matrix:
+    runs-on: ubuntu-latest
+    outputs:
+      apis: ${{ steps.set-matrix.outputs.apis }}
+    steps:
+    - uses: actions/checkout@v4
+    - id: set-matrix
+      run: |
+        apis=$(jq -c '[.[].version]' apicompat/apis.json)
+        echo "apis=$apis" >> $GITHUB_OUTPUT
+
+  nuget-package:
+    runs-on: ubuntu-latest
+    needs: generate-matrix
+    strategy:
+      matrix:
+        api: ${{ fromJson(needs.generate-matrix.outputs.apis) }}
+      fail-fast: false
+
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: Setup .NET
+      uses: actions/setup-dotnet@v4
+      with:
+        dotnet-version: '8.0.x'
+
+    - name: Install Microsoft.DotNet.ApiCompat.Tool
+      run: dotnet tool install --global Microsoft.DotNet.ApiCompat.Tool
+
+    - name: Restore dependencies
+      run: dotnet restore
+
+    - name: Make script executable
+      run: chmod +x apicompat/test-api-compatibility.sh
+
+    - name: Run API Compatibility Check for ${{ matrix.api }}
+      run: ./apicompat/test-api-compatibility.sh ${{ matrix.api }}
