+++ apicompat/apis.json
@@ -1,0 +1,78 @@
+[
+  {
+    "version": "6.4.6",
+    "release_date": "2025-09-22"
+  },
+  {
+    "version": "6.4.2",
+    "release_date": "2025-05-12"
+  },
+  {
+    "version": "6.4.1",
+    "release_date": "2025-05-12"
+  },
+  {
+    "version": "6.4.0",
+    "release_date": "2025-04-03"
+  },
+  {
+    "version": "6.3.1",
+    "release_date": "2024-01-03"
+  },
+  {
+    "version": "6.3.0",
+    "release_date": "2023-10-20"
+  },
+  {
+    "version": "6.2.6",
+    "release_date": "2023-06-12"
+  },
+  {
+    "version": "6.2.5",
+    "release_date": "2023-05-10"
+  },
+  {
+    "version": "6.2.4",
+    "release_date": "2023-04-27"
+  },
+  {
+    "version": "6.2.3",
+    "release_date": "2023-01-18"
+  },
+  {
+    "version": "6.2.2",
+    "release_date": "2023-01-06"
+  },
+  {
+    "version": "6.2.1",
+    "release_date": "2023-01-06"
+  },
+  {
+    "version": "6.2.0",
+    "release_date": "2022-05-02"
+  },
+  {
+    "version": "6.1.0",
+    "release_date": "2021-11-02"
+  },
+  {
+    "version": "6.0.8",
+    "release_date": "2021-08-09"
+  },
+  {
+    "version": "6.0.7",
+    "release_date": "2021-06-15"
+  },
+  {
+    "version": "6.0.6",
+    "release_date": "2020-11-12"
+  },
+  {
+    "version": "6.0.5",
+    "release_date": "2020-11-10"
+  },
+  {
+    "version": "6.0.3",
+    "release_date": "2020-10-30"
+  }
+]
